Using wrong $arch on ppc64le
On both ppc64 and ppc64le, the same ppc64_defaults is used.
But the $arch is assumed to be embedded as part of the
catalog entry which means that ppc64le changes to ppc64 in
callBootMethods subroutine.

This fix just adds a ppc64le method that call the
ppc64 method with the same parameters and correct $arch.