Add special RPi hybrid mode
One of the problems with the Raspberry Pi is that it does not contain
smart firmware - it relies on the SD card or USB storage to provide
most of the firmware smartness.

That means in order to make the iso file bootable directly for installation,
we need to provide that firmware on the ISO in a bootrom readable way.

So far that means we need provide an MBR partition table with a VFAT
partition inside that has the partition type 0xc (FAT32 LBA).

This patch adds support to create such a hybrid MBR based on the existing
hybrid MBR support. The only difference is that instead of relying on the
isohybrid tool (which also puts x86 code into the iso), we just assemble
our own MBR straight away.

Signed-off-by: Alexander Graf <agraf@suse.de>