fix iso creation (bsc#1162531)
mkisofs was not called with exactly the same parameters in the first and
second run. But it is essential that the generated isofs meta data are of
the same size in both runs.

The old code left a superfluous 'glump' directory entry in the first run
which could lead to a potential block-misalignment.