post process build output to reduce rsync disk io
by using the kiwi_post_run hook to set the mtime of files that are the
same as the last build to the last mtime.

If rsync is used without its --checksum argument it would skip files in
error if they do not have a newer mtime. But then always using new
mtimes would unecessarily use disk io to be able to skip transferring
the content. This change optimizes that.

Ideally we would instead fix mirroring to consider also the file content
even when the meta data is equal, either by using a sync solution that
can keep that as state to not need to rehash files all the times or by
accepting the disk IO cost. If we ever do we can remove this change
again and set the mtime of all build output to SOURCE_DATE_EPOCH.