Introduce DOWNLOAD_MIRROR_POLICY option
It will lead to a build failure, when an rpm with the same name,
but different scheduler architecture exists in multiple versions.

This can happen for example with noarch packages which fail on
some scheduler architectures or when doing tricks with exportfilter
and aggregates.

This avoids that a package content may change on the same location,
where some download mirrors or redirectors have a problem to identify
the file.

We do not enforce a certain scheduler arch here on purpose since this
can lead to dependency issues when arch packages may require an
exact subpackage version-release with unique (noarch) architecture.